数据库bigint改为映射成Long

+++ src/main/java/xyz/fz/api/Order.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Date;
 
 @Data
 public class Order implements Serializable {
-    private BigInteger orderId;
+    private Long orderId;
 
-    private BigInteger userId;
+    private Long userId;
 
     @JsonFormat(timezone = "GMT+8", pattern = "yyyy-MM-dd HH:mm")
     private Date createTime;
