+++ src/main/java/xyz/fz/controller/OrderController.java
@@ -6,7 +6,6 @@ import xyz.fz.api.Order;
 import xyz.fz.service.OrderService;
 
 import javax.annotation.Resource;
-import java.math.BigInteger;
 import java.util.*;
 
 @RestController
@@ -19,14 +18,14 @@ public class OrderController {
     @RequestMapping("/create")
     public String create() {
         Random random = new Random(9999L);
-        orderService.create(BigInteger.valueOf(System.currentTimeMillis()), random.nextInt(100), random.nextInt(10));
+        orderService.create(System.currentTimeMillis(), random.nextInt(100), random.nextInt(10));
         return "create ok";
     }
 
     @RequestMapping("/createAtTwoDs")
     public String createAtTwoDs() {
         Random random = new Random(9999L);
-        orderService.createAtTwoDs(BigInteger.valueOf(System.currentTimeMillis()), random.nextInt(100), random.nextInt(10));
+        orderService.createAtTwoDs(System.currentTimeMillis(), random.nextInt(100), random.nextInt(10));
         return "create at two ds ok";
     }
 
@@ -37,7 +36,7 @@ public class OrderController {
         list.add(randomProduct());
         list.add(randomProduct());
         list.add(randomProduct());
-        orderService.batchOrder(BigInteger.valueOf(System.currentTimeMillis()), list);
+        orderService.batchOrder(System.currentTimeMillis(), list);
         return "batch create ok";
     }
