+++ src/main/java/xyz/fz/entity/OrderItemEntity.java
@@ -3,7 +3,6 @@ package xyz.fz.entity;
 import xyz.fz.api.OrderItem;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 
 @Entity
 @Table(name = "t_order_item")
@@ -12,19 +11,19 @@ public final class OrderItemEntity extends OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_item_id")
     @Override
-    public BigInteger getOrderItemId() {
+    public Long getOrderItemId() {
         return super.getOrderItemId();
     }
 
     @Column(name = "order_id")
     @Override
-    public BigInteger getOrderId() {
+    public Long getOrderId() {
         return super.getOrderId();
     }
 
     @Column(name = "user_id")
     @Override
-    public BigInteger getUserId() {
+    public Long getUserId() {
         return super.getUserId();
     }
