+++ src/main/java/xyz/fz/service/OrderService.java
@@ -2,16 +2,15 @@ package xyz.fz.service;
 
 import xyz.fz.api.Order;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
 public interface OrderService {
-    void create(BigInteger userId, Integer productId, Integer productCount);
+    void create(Long userId, Integer productId, Integer productCount);
 
-    void createAtTwoDs(BigInteger userId, Integer productId, Integer productCount);
+    void createAtTwoDs(Long userId, Integer productId, Integer productCount);
 
-    void batchOrder(BigInteger userId, List<Map> products);
+    void batchOrder(Long userId, List<Map> products);
 
     List<Order> listLast10();
 }
